requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.0.0
ultralytics>=8.2.0
Pillow>=10.0.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
openpyxl>=3.1.0